+++ src/main/java/com/travel/role/global/exception/GlobalExceptionHandler.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import com.amazonaws.SdkClientException;
 import com.travel.role.global.exception.comment.CommentInfoNotFoundException;
+import com.travel.role.global.exception.common.ResourceOperationAccessDeniedException;
+import com.travel.role.global.exception.common.S3ImageNotFoundException;
 import com.travel.role.global.exception.room.AlreadyExistInRoomException;
 import com.travel.role.global.exception.room.InvalidInviteCode;
 import com.travel.role.global.exception.room.InvalidLocalDateException;
